Change from one domain = default to external domain search
Before, if you only had one domain defined, it would be used for
all requests no matter the hostname.

Now, all requests must match an explicitely configured domain name.
To help, we now have access to a DB of Domain to Module mappings.
You can populate the configuration DB with other hosts to be loaded
on-demand (including having the DB hosts specify their module as a
binary to load directly instead of from a filename).

The next step in improving this feature will be making the DB lookup
more extensible (move to Redis from ETS or make the DB fetcher pluggable).